bashrc: tmux: Match target sessions exactly
See https://github.com/tmux/tmux/issues/346.

From the tmux man page:

target-session is tried as, in order:

  1.   A session ID prefixed with a $.

  2.   An exact name of a session (as listed by the list-sessions
       command).

  3.   The start of a session name, for example `mysess' would match a
       session named `mysession'.

  4.   An fnmatch(3) pattern which is matched against the session name.

If the session name is prefixed with an `=', only an exact match is
accepted (so `=mysess' will only match exactly `mysess', not
`mysession').

If a single session is found, it is used as the target session; multiple
matches produce an error.  If a session is omitted, the current session
is used if available; if no cur- rent session is available, the most
recently used is chosen.