vimrc: Simplify SerbanDeleteBuffer()
Same behavior as :bd but keep the current window when closing a
non-empty buffer.

This preserves the splits and window layout to a first approximation,
but unfortunately if the buffer is open in other windows besides the
current one, :bd closes those other windows. A nice improvement could be
to retrieve all the windows containing the buffer and run :enew on them
as well. This would preserve splits and window layouts everywhere.

This could be done with win_findbuf() and win_execute():

  for id in win_findbuf(bufnr())
    call win_execute(id, 'enew')
  endfor

SerbanDeleteEmptyBuffers() would need to be tweaked to make this work.