vimrc: Synchronize text and comment wrapping
When manually changing the wrap mode for text files, wrapping of normal
text and wrapping of comments (as specified by the `comments` option)
should be in sync so that paragraph text and list text receive the same
treatment. This is in contrast to, say, Go, in which we want to wrap
comments but not code. In that case, `formatoptions` would contain 'c'
but not 't'.