less: Remove key binding for saving input to file
By default, pressing 's' presents the `log file:` prompt. This binding
is cryptically defined as `s toggle-option o` in the lesskey man page
since passing `-o` on the command line "causes less to copy its input to
the named file as it is being viewed", whereas the built-in help calls
this binding "save input to a file".