vimrc: SerbanOpenLink(): Add custom link opener
Support opening http and https URLs, short links, and usernames.

Once again, vim-markdown interferes and breaks things. It remaps gx to
<Plug>Markdown_OpenUrlUnderCursor / s:OpenUrlUnderCursor() which yields
'The cursor is not on a link' for reasons I don't care to ascertain.

netrw can also open links, but to support custom link formats it's
easier to circumvent netrw entirely.