tmux.conf: Set the terminal clipboard using OSC 52
This change makes tmux emit OSC 52 to the terminal both when making a
selection using copy mode and when an application running inside of tmux
itself emits OSC 52.

The wording in the man page for `set-clipboard` is confusing. It says:

› If set to `on`, tmux will both accept the escape sequence to create a
› buffer and attempt to set the terminal clipboard. If set to
› `external`, tmux will attempt to set the terminal clipboard but ignore
› attempts by applications to set tmux buffers.

The wiki page clarifies this:
↳ https://github.com/tmux/tmux/wiki/Clipboard#changing-set-clipboard

› The difference is that `on` both makes tmux set the clipboard for the
› outside terminal and allows applications inside tmux to set tmux's
› clipboard (adding a paste buffer). `external` only makes tmux set the
› clipboard and forbids applications inside from doing so.

See also https://github.com/tmux/tmux/issues/1864:
↳ terminal clipboard not set when set-clipboard is 'external'