vimrc: nvim: Hard-code the Python interpreter path
Neovim relies on the pynvim Python package to run Python plugins.
That package isn't available inside of a virtualenv, so point Neovim to
the system Python so that it searches for pynvim in the site packages
directory.

Also, the Python interpreter Neovim runs should generally not depend on
the virtualenv Neovim is running in.

See `:help provider-python`.