fish: g: Print file names without quoting
This is the default behavior of GNU ls when piping the output into
another program. When g is called with the `--no-pager` option, the
output is not piped into less, resulting in different formatting.

Pass the `--literal` flag to GNU ls in order to disable quoting always.

This is a follow-up to 7e3b195965a2f3bf34dec68f9cdbdfefea6d7b16.