vimrc: markdown: Restrict link title matching
Both SerbanLinkSeparator and SerbanLinkTitle need to be marked as
`contained` so that naked hyphen sequences used for bulleted lists
(` - `) are not recognized as a SerbanLinkSeparator followed by a
SerbanLinkTitle.

Exclude U+2502 (`│`) from the SerbanLinkTitle pattern in order to
contain link title syntax highlighting within a foxytables cell.

This is a follow-up to 05dee3d1a50617e163afd5296d80f1c33d26ab7e.