vimrc: Disable gdefault
When calling :substitute, do not replace all occurrences on a line by
default.

Setting `gdefault` makes it so that calling `s/old/new/` is like calling
`s/old/new/g`. That would be great on its own, which is why I enabled it
in the first place, but it also INVERTS THE MEANING of the `g` flag:
instead of replacing all occurrences of a match, it only replaces the
first occurrence as if `gdefault` was off and the `g` flag was omitted.

This behavior is unintuitive and unexpected, and it breaks plugins.

From `:help :s_g`:

  If the 'gdefault' option is on, this flag is on by default and the [g]
  argument switches it off.

From `:help :gdefault`:

  DEPRECATED: Setting this option may break plugins that are not aware
  of this option.  Also, many users get confused that adding the /g flag
  has the opposite effect of what it normally does.

See:

• :help 'gdefault'
• :help :s_flags
• :help :s_g
• https://github.com/vim/vim/issues/1767
  ↳ Behaviour change suggestion: /g substitute always global even if
    gdefault set