script.py: run(): Print correct exit status
The existing code in run() has a minor bug:

    if returncode := _run_and_indent_output(args) != 0:
      raise subprocess.CalledProcessError(returncode, args)

The intended value of `returncode` is the result of
`_run_and_indent_output(args)`, an integer, but the actual value is the
result of the expression `_run_and_indent_output(args) != 0`, a boolean.
When the process generates a nonzero exit status, the program flow still
ends up being correct, but the value of `returncode` is `True` instead
of the integer exit status. This went undetected because
`CalledProcessError.__str__()` produces the “Command … returned non-zero
exit status 1” error message using the integer format specifier `%d` for
`returncode`. Because 1 is a common failure exit status and `int(True)`
evaluates to 1, the output was often correct for the wrong reason.

It's more Pythonic to omit the explicit comparison to zero entirely:

    if returncode := _run_and_indent_output(args):
      raise subprocess.CalledProcessError(returncode, args)

But I prefer to keep it as a reminder that `returncode` is supposed to
be an integer exit status. The walrus operator has the lowest
precedence, so parentheses must be used to assign the correct value to
`returncode`:

    if (returncode := _run_and_indent_output(args)) != 0:
      raise subprocess.CalledProcessError(returncode, args)

See:

• https://docs.python.org/3/reference/expressions.html#assignment-expressions
• https://docs.python.org/3/reference/expressions.html#operator-precedence