fish: config: FZF_DEFAULT_OPTS: Make check robust
The Homebrew build of fzf reports its version as “0.66.0 (brew)”, but
the Debian build of fzf reports its version as “0.60 (devel)”, without
the patch number. In the latter case, splitting on `.` does not result
in a valid integer:

› Integer 60 in '60 (devel)' followed by non-digit
› ~/.config/fish/config.fish (line 154):
› test (fzf --version | cut -d . -f 2) -ge 66
› ^
› from sourcing file ~/.config/fish/config.fish

Use a regular expression to extract the minor version number instead.

This is a follow-up to commits:

• f9307f6270f6da141bd946d318b0e52ab853b08c (2025-10-21)
  ↳ fish: config: FZF_DEFAULT_OPTS: v0.66: Hide gutter